Parametrizable plugin fields for target generators (#20059)
Provides the machinery to be able to fix the issue in #19186 

## Problem

We could not previously do all three:
- target generator
- parametrize 
- plugin field
Which broke use-cases like protobuf code generation from a
`protobuf_sources` target with multiple python resolves.

See first part of [comment
below](https://github.com/pantsbuild/pants/pull/20059#issuecomment-1771572260)
below for description of the technicalities of the issue.

## Implemented solution

Follows [this
proposal](https://github.com/pantsbuild/pants/pull/20059#issuecomment-1771603843).
We extend the `register_plugin_field` method on `TargetGenerator` with a
parameter `as_moved_field`:
```
  ProtobufSourcesGeneratorTarget.register_plugin_field(
      ProtobufPythonResolveField, as_moved_field=True
  ),
```
When a plugin field is registered with `as_moved_field`:
- It will behave for the purposes of parametrization and generation, as
though it was one of the `moved_fields` on the original generator,
rather than a `copied_field`. This makes it parametrizable.
- In other respects the field should behave as before.

We considered making `as_moved_field` the default behavior, since it is
likely what is intended by plugin authors, but we did not. We can make
that larger-surface-area change later.

## Testing

See follow-up PR https://github.com/pantsbuild/pants/pull/20069 for E2E
testing against an example repo