Add additional metrics for remote byte stores (#20138)
This adds a bunch of additional metrics to provide more insight into the
byte store behaviour: for each of the "read", "write" and "list missing
digests" (aka "exists") operations, track:

- number of attempts
- number of successes (for "read", this is further split into "cached"
and "uncached", i.e. whether the successful network request found data
or not)
- number of errors

This mimics the metrics for the remote action cache.

The minor shuffling of code in this PR also fixes a bug with the
`RemoteStoreBlobBytesDownloaded` metric: before this PR, any successful
network request would count as bytes downloaded, even if the digest
didn't exist/wasn't downloaded (i.e. it successfully determined that the
digest wasn't cached). This PR restricts the metric to only count digest
bytes when the network request is successful _and_ the digest actually
existed/was downloaded.

I'm hopeful this will at least give more insight into the scale of the
problem in #20133, in addition to being generally useful as "what's
pants up to" reporting.