Python: Enable code formatting using Ruff (#20098)
Fixes #20084.

**Warning**: This will enable the Ruff formatter by default for all
repos that use the Ruff backend for linting.

To work around this and run formatting using Black instead of Ruff, you
can use the `only` option to the `pants fmt` command like so:

`pants fmt --only=black` 

If using another formatter instead of Black, substitute the name of the
tool in the command above.

Additionally, if you'd like to skip Ruff entirely, you can use the
[`skip`](https://www.pantsbuild.org/docs/reference-ruff#skip) option
defined in the Ruff tool config. Note that this will skip **all three**
of formatting, linting, and fixing.