Remove documentation about new remote caching in 2.19 (#20183)
This takes the unusual step of _removing_ documentation. In particular,
this removes the documentation for the still _very_ experimental new
remote cache backends: GitHub Actions and file system, because they're
useless and very unstable:

- there's significant user-facing refactoring work that I'm not going to
get to for 2.19 (https://github.com/pantsbuild/pants/issues/19902),
particularly around changing how the options are configured
- the GHA backend seems to be very... limited, and only works for small
repos (https://github.com/pantsbuild/pants/issues/20133)

I'm targeting 2.19, as it stabilises, and leaving the docs in `main`
(for 2.20).