Support inferring JS deps from `export ... from ...` statements (#20190)
This extends the dependency inference for JS to also consider `export`
statements. This uses all the cases from
https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Statements/export
as test cases.

This also adds some tests for how `// pants: no-infer-dep` is
interpreted for multi-line statements for `export`, and also for the
existing support for `import ...`, `require(...)` and `import(...)`.
This may not be the behaviour we exactly want, but we at least now don't
change it accidentally. #20207 covers nailing it down.

Fixes #19819