Fix spurious warnings about Helm interpolations due to incorrect comparison (#20220)
This fixes the second half of #20210 by more accurately detecting when
the helm `values` value had deprecated `{env.X}`-style interpolations
applied. Previously, it compared a `FrozenDict` (`current_value`) to a
`dict`, which always returns non-equal (#20219, a very subtle issue),
i.e. implying that there was changes, even when there was not.

This rewrites the code to avoid comparing `FrozenDict`s because that's
currently so "dangerous".

(#20217 fixed the first half.)