Allow passing arbitrary args to PEX invocation when building FaaS artifacts (Cherry-pick of #20237) (#20281)
This adds an `pex3_venv_create_extra_args` field to all FaaS targets
(`python_aws_lambda_function`, `python_aws_lambda_layer`,
`python_google_cloud_function`). This allows adding arbitrary extra
arguments that are passed to the `pex3 venv create --layout=flat-zipped
...` invocation that is used to create the final zip file.

Most acutely, this is driven by making it possible to pass the
`--collisions-ok` flag. This allows work around dependencies that are
packaged with files outside a namespaced directories, e.g. commonly a
LICENCE or README file, or `tests/` directory, since they'll have
different content. A command like `pip install ...` will happily install
them and have one of the files "win" arbitrarily, while PEX is more
correct and flags that it doesn't know what to do in that circumstance.

Fixes #20224

This is marked for cherry picking back to 2.18 because it can block
adoption of the new layout, and the old (lambdex) layout is deprecated
and using it is noisy in 2.18.