generate lockfiles based on `terraform_module`s
It doesn't uphold the idea that `terraform_module`s are child modules  and `terraform_deployment`s are root modules.
This distinction was shown to be unintuitive both by feedback with real users
and by how difficult it was to explain how lockfiles interacted with multiple deployments and why lockfiles were targeted by deployments but appeared under modules.