Change the slug of the Shell backend doc.
It used to be "shell", now it's "shell-overview" because we
were getting 401s from readme.com for that page, and their
support claims we discovered a bug whereby "shell" is not
a valid slug.

This will all be moot once we switch off readme, but this is
a bandaid for now.