Fix algorithm for gathering Go build requests for cgo.
The previous algorithm did not check if a package had already been
traversed, which leads to exponential blowup of the queue.

Go doesn't support dependency cycles, so the previous algorithm
would converge eventually, but before it did so the queue could get
to sizes that were infinite in practice. This happened in a real-world
case.

This is an identical fix to the one in #20030, which was for
coverage build requests. These are the only two places in the go
backend (that I can find) where a transitive walk like this is done.