Fix algorithm for gathering prebuilt Go object files (#20332)
The previous algorithm did not check if a package had already been
traversed, which can lead to exponential blowup of the queue.

This is an identical fix to the one in #20030, which was for coverage
build requests. These are the only two places in the go backend 
(that I can find) where a transitive walk like this is done.

Unlike #20030 this one is not known to have caused a problem for any
users in practice.