Support bool, float and lists of them in the rust options parser. (#20388)
We don't document on the docsite that we support float (and list of
float) but it is supported in the code, and we do actually have one
float option in the codebase. So it's possible more exist in plugins,
and therefore we have to  support it.

"List of bool" seems like a silly option type, and we have no such
options, but alas we do support it and document that fact, so again,
we must  support it. As a byproduct of supporting list of bool we now
parse bool in the PEG parser and so can remove the previous bespoke
bool parsing function.

This adds a lot of tests for parsing scalars and lists of them, and also
tests individual quoted string parsing. But it does not slavishly
perform every test on every scalar type. In some important cases it
tests bool/int/float lists separately, but in others it uses one or the
other as a stand-in for all scalars.

A future change will support dicts with values of the above.