python-protobuf backend now support multiple protoc plugins. (#20387)
This change adds support for grpclib in addition to the existing grpcio
support.

This PR does not change any default behavior. User are able to opt-in to
the new plugin with `--python-protobuf-grpclib-plugin` and opt-out of
the past/current default with `--no-python-protobuf-grpcio-plugin`. It
should be possible to easily add support for `python-betterproto` plugin
as well.

See https://github.com/pantsbuild/pants/issues/20383