Terraform lockfiles (take 2) (#20303)
Generate and consume Terraform lockfiles! Each `terraform_module` can
have a lockfile generated for it. The lockfiles use the target address
of the module as the resolve name.

Using `terraform_module`s here instead of `terraform_deployment`s
doesn't uphold the idea that `terraform_module`s are child modules and
`terraform_deployment`s are root modules. This distinction was shown to
be unintuitive both by feedback with real users and by how difficult it
was to explain how lockfiles interacted with multiple deployments and
why lockfiles were targeted by deployments but appeared under modules.

This MR also reworks how Terraform pulls in backend configs and var
files. These used to be specified simply as files, which was kludgy.
(Multiple different sources fields meant requests for SourceFiles didn't
work, since they would only pull in one of those fields.) Talking it
over, we determined that the best design would be to have separate
targets for `terraform_var_files` and `terraform_backend`, each with
their own sources field, and pull these in with dep inference. We could
ease some of the boilerplate by automatically inferring a dep on these
targets in the same directory as the `terraform_deployment`. This can be
overridden by manually specifying these in the `dependencies` field.
This feature was bundled in this MR because it is linked to the roles
that deployments and modules have in Pants's understanding of Terraform,
which is an issue precipitated by lockfiles. That is, design decisions
we need to make to implement lockfiles will also require changes in
these targets.

Supported features:
- Create Terraform lockfiles for each module
- Use Terraform lockfiles for deployments, check, &c.
- Automatic dependency inference on backend config and var files in the
deployment's directory

Limitations:
- Multiple deployments sharing the same root module share the same
lockfile (this is how it works in Terraform)

This is the second pass at this. The
[first](https://github.com/pantsbuild/pants/pull/19762) was getting
cluttered by design changes.