Allow AbstractLintRequest subclasses to disable lint rules (#20407)
By default any fixer in pants has builtin linter, that runs the fixer
and checks if it did any changes, but sometimes it needs to be disabled.
For example, ruff is both a linter and a fixer, and the fixer can only
fix a subset of what linter shows as an error.
This means there is no need to enable "auto"-linter for the fixer,
because it is already covered by `ruff check` linter. This pr introduces
a flag `enable_lint_rules`, that allows you to disable builtin fixer
linter.