Fix direct_url in python PEP660 editable wheels
We were using the first entry in `build_time_source_roots` to populate
the editable wheel's `direct_url`. However, that does not work as I
expected it to because `build_time_source_roots` gets sorted before
creating the `DistBuildRequest`, so the first entry is not actually
the `source_root` of the `python_distribution` target.

To fix this, we capture the dist's source_root before the source_roots
get sorted and use that when building the PEP660 editable wheel.