Fix direct_url in python PEP660 editable wheels (Cherry-pick of #20486) (#20494)
We were using the first entry in `build_time_source_roots` to populate
the editable wheel's `direct_url`. However, that does not work as I
expected it to because `build_time_source_roots` gets sorted before
creating the `DistBuildRequest`, so the first entry is not actually the
`source_root` of the `python_distribution` target.

To fix this, we capture the dist's source_root before the source_roots
get sorted and use that when building the PEP660 editable wheel.

Functionally, the editable wheels get installed just fine. However, the
output of `pip list` in an exported venv might show the wrong path if
there happens to be a source_root that sorted before the source_root of
the distribution. Importing the python code still works, however, as all
of the source_roots are included in the `.pth` file. Thus, this is a
cosmetic / correctness bug.

Co-authored-by: Jacob Floyd <cognifloyd@gmail.com>