Use Python subsystem's default ICs when generating builtin lockfile (#20500)
This eliminates a bit of duplication of ICs between the script for
generating lockfiles for builtin tools, and the subsystems for those
tools. In particular, subsystems (optionally) have a class property
indicating their interpreter constraints, which previously had to be
then duplicated into the script for generating lockfiles so that the
lockfile used those constraints rather than the system-wide default
`>=3.7,<4` ones.

This affected only the `pytype` subsystem, which happened to actually
have inconsistent interpreter constraints (3.10.* was excluded in the
subsystem, but allowed by the locker). I've made them match what the
lockfile was being generated with here, I'm not 100% sure this is
correct.

(This is a refactoring I noticed when trying to do #20501, since that
requires adjusting the lower-bound of the semgrep ICs and it wasn't
working as I expected, due to missing this duplication.)