Handle unresolved ambiguous entrypoint dependency for PEX as unowned dependency (#20390)
Pants will infer a dependency for a pex_binary on the file referenced in
the entrypoint field. An ambiguous entrypoint for PEX would result in us
inferring no depedency. This is a problem when an added file results in
an ambiguity, as now a PEX that used to contain a file no longer does.
This MR creates a special case for when we find an ambiguity for the
entrypoint's file and are unable to resolve it through explicitly
provided dependencies. I decided that tying this dependency inference
operation into the `unowned_dependency_behavior` would make the most
sense. It allows users the option to cause this to error or to preserve
the existing behaviour. I thought of always throwing the error, which
I'm pretty sure is correct; but that would not give the option to
disable that.
Note that if no owners are found (both non/ambiguous) we simply return
`()`, as before