docs: provide example how to convert target addresses to source files in rules API (#20524)
A plugin may need to output a list of resources (e.g. source code files)
so that it can be fed to another program (outside of Pants). Pants
operates on targets, and they can won't be understood by anything
outside of Pants.

E.g. `some/config/folder/example1.cfg:../config` would need to be
trimmed to map to a file. This is similar to what `filedeps` goal would
do, but without including BUILD files which are often irrelevant.

This snippet shows how this is done.

---------

Co-authored-by: Andreas Stenius <git@astekk.se>