Generate option value derivations in the rust parser. (#20535)
We use these to show the source of option values in help output.

The derivations can be a lot of data, so we only produce them
if asked to.

Note that the derivations currently only show the source type
(args, env, config, default) but not, for example, which config
file a value came from. A future change may add this,  but it
will require some changes to the Source enum which are out
of scope for this PR. 

Also: use maplit for easy hashmap literals in the test.