stats: add output_file option to output the stats to a file (#20512)
Currently when running a Pants goal with the `--stats-log` enabled, the
stats are showing at the end of the goal's output. This makes
distinguishing between the goal output and stats difficult. One cannot
redirect only the stats data to a file as it's just part of the goal
output.

This PR adds a new option to request saving the stats into a file to be
able to parse it later or archive.

As a next step, I'd like to see a structured data produced. JSON seems
sensible so that one can mung it to obtain data in any desired format
such as [Prometheus](https://prometheus.io/docs/practices/naming/). The
logged stats data is not really suitable for parsing.

Co-authored-by: Huon Wilson <wilson.huon@gmail.com>