Inline multiple config files into options sources. (#20549)
Previously the options sources list could have only one source
per type (flag/env/config/default). Since we support multiple
config files, we had the concept of a "merged config" that
encompassed multiple "config sources". This led to
duplication of the logic that traverses and merges sources.

This change gets rid of the concept of merged config. 
Instead, the list of options sources we consult can now
include multiple config files. This is achieved by enhancing
the Source::Config enum to include an ordinal (to ensure 
that configs are applied in the right order) and the path to
the config file (purely informational). 

The natural `#[derive(Ord)]` will order these augmented
configs appropriately within all the sources. To make this
easier to achieve we change the order of the enum to be
from lowest to highest priority, so that we can add config
files naturally in the order in which we encounter them.

 This also allows us to simplify some types. In particular, 
we no longer need to traffic in `Vec<DictEdit>` - a single
source can only return a single `DictEdit`.

More importantly, this supports derivation history that
includes the specific config file each value comes from.