Support @fromfile option values in the Rust options parser. (#20634)
As part of this, streamlines away some previous function pointer
clunkiness via a new `Parseable` trait that implements parsing
of scalars and list of scalars for the scalar types bool/i64/f64/String.

While messing with this code: this change also implements the 
"source" field for list- and dict-valued option values.
Previously this was only provided for scalar-valued options. However
the highest-priority source is still a valid and useful concept even
for compound values that might be composed across multiple sources.