Remove force flags from release tagging & confirm commit before doing so (#20640)
This does two changes in response to my mistake with retagging
2.21.0.dev0
(https://pantsbuild.slack.com/archives/C0D7TNJHL/p1709588198420159):

- adds a check that the commit-to-be-tagged is correct, via `git show`
- removes the `-f` (force) flags from both the `git tag` and `git push`
invocation, so that if a tag already exists, we're not automatically
overwriting it.
 
The second of these is explicitly revisiting #4504 (that introduced the
flags). I believe our release process is quite different to how it was
then, and thus we can be more careful with tags now. And, if someone
really needs to move a tag, they can always add the `-f` flags
themselves, as required (or delete the tag, or similar). This is
referenced in a comment.