Fix the exception sink test error message check. (#20655)
On my M1 I get a different message than apparently CI does.

So now we allow one of a set of messages to match, per platform.

Note that this error message check was already fragile, and the
exact nature of when we get each message is not understood, 
as admitted by a comment there. This PR doesn't address
the underlying problems with this test, but just gets things
working again without breaking things in CI or on other
platforms.