Tighten up options tests (#20660)
This fixes a few loose behaviors in the tests that happen 
to work, but that we have never documented or claimed
to support.

Notably, we test an option type of "list of dicts", but we
have no actual options of that type, have never documented
support for it, and it is not supported by our new-style
declarative registration classes.

We also tighten up a couple of places where we implicitly
relied on the assumption that option values (including
list members) were roundtripped via a `str`. But of 
course this is not guaranteed, and we have never claimed
that this should work.

These were noticed while manually testing the Rust
options parser against these Python tests. This PR is 
groundwork for switching Python code to use the
rust parser.