`complete` built-in goal which can be called by completion scripts (or IDEs) to process a list of commands (#19478)
This commit adds a new `complete` built-in goal which facilitates "repo-centric" shell completions.

The first-run is inherently slow, and then it is much quicker after the first load. The annoying caveat at the moment is that the filesystem can only be accessed with a relative, absolute, or tilde specifier - in order to short-circuit calling pants, and using zsh/bash functionality instead.

pants complete --shell=bash > pants-completions.bash
pants complete --shell=zsh > pants-completions.zsh

That file needs to be appropriately sourced via zsh, or bash, and whatever extra `compinit` or `autoload` as per the developer's system.