use default upstream and upgrade shellcheck (#20702)
Now that the default upstream produces `darwin.aarch64` builds, we can
switch to the regular upstream instead of the vscode fork.

Note that the vanilla upstream puts the version number in the archive
name, so `generate_exe` is now compatible with vanilla instead of the
forks norms. This means "downgrades" <0.10 will not work. This happened
to be how the example at
`docs/docs/writing-plugins/the-rules-api/installing-tools.mdx` already
worked.

docs/docs/writing-plugins/the-rules-api/installing-tools.mdx-

Funny Changelog:
 * https://github.com/koalaman/shellcheck/releases/tag/v0.10.0
 * https://github.com/koalaman/shellcheck/releases/tag/v0.9.0

Actual changelog:
https://github.com/koalaman/shellcheck/blob/master/CHANGELOG.md#v0100---2024-03-07