update k8s_parser lock; broaden model compatibility
In v1.1.1 hikaru switched from one big package, to a collection of
packages:
https://github.com/haxsaw/hikaru/blob/main/release_notes.rst#v110

This means that 'which version of hikaru are we using?' and 'which k8s
model versions do we support?' become somewhat separte.  The sets the
minimum version to be where the package revamp happend, and includes
the broadest set of model versions currently available.