Add workflow to clear persistent cache dirs on self-hosted runners (#20755)
We have a few long-lived runners that have cache directories that grow
and grow until the machine runs out of space. This workflow is a
stop-gap measure that makes it easier for anyone to give the machine
more space, by making workflow to do it, rather than requiring someone
with credentials to SSH in and do it manually.

For instance, on March 18, the macOS 11 runner had:

| directory              | size (GB) |
|------------------------|-----------|
| `~/Library/Caches/nce` | 113       |
| `~/.cache`             | 34        |
| `~/.rustup`            | 12        |
| `~/.pex`               | 8.8       |
| `~/.nce`               | 1.2       |

This PR set-ups automatic deletion of these directories with some
logging.