Use correct GHA workflow syntax in new cache-clearing action, be more resilient (#20756)
In #20755, I had a typo: formatting the jobs as a list, with `name`s
keys... but that's the syntax for steps. Jobs need to be a dict with the
ID as the key.

This also squashes any errors from `rm -rf ....`: on Mac, it seems that
attempts to delete directories that the runner doesn't have permission
for (particularly system ones in `~/Library/Caches`). But that's fine,
we'll still delete all of the caches created by "day-to-day" actions
runs, which are the big ones.

(Why wasn't this caught in #20755? I attempted to run the workflow
before merging, and got errors, but incorrectly assumed this was because
the workflow didn't exist on `main`, but they were actually because it's
invalid.)

This has now successfully run and cleared hundreds of gigs over two
runs:

- https://github.com/pantsbuild/pants/actions/runs/8562624080 (required
the "squash `rm` errors fix)
-
https://github.com/pantsbuild/pants/actions/runs/8562947394/job/23467155450
(with that fix).