Deprecate the implicit_value option registration kwarg. (#20757)
It was never used in the Pants repo, and was only implemented back in
2015, as part of removing our use of argparse, for feature parity.

It is undocumented, and not supported by the newer BoolOption
registration mechanism. So the only place this could be used is
in old-style option registration in a third-party plugin, which seems
pretty unlikely.

However, for good practice, we give a short deprecation period, just in
case.