stats: add output format option to save Pants stats as JSON (#20579)
With the current implementation, when stats output is shown, it's always
in plain text which makes parsing the stats really hard. For instance,
it may be useful to track the cache hit rate for every Pants invocation
to make sure it's being used effectively.

To make parsing the stats easy, it's now possible to export stats as
JSON Lines objects. The amount and type of information is identical to
what's being produced for plain text output, only the output format
differs.

A few implementation details worth sharing:
* when saving the output into a file (that may already have previous
executions saved), the write happens in append mode to keep the stats of
a previous execution. This makes it possible to have a single place with
all the Pants invocations stats (e.g. after a CI build where multiple
separate Pants invocations might have happened is complete).

As a follow-up, I'll refactor the code responsible for production of the
stats so that the logic is defined in a single place (e.g. getting all
the data in JSON format and then dumping it into plain text, if user
asks to do so).