refactor nodes module into submodules by kind of node (#20742)
Refactor the `nodes` module in the `engine` crate into a submodule for
each kind of node. Each node type is re-exported by the `nodes` module.

Notes:
- Code in `NodeKey` (and other parts of `nodes` module) reaches into
what should be private fields of each node. I have marked these fields
as `pub(super)` for now since I want this PR to be mostly a pure source
rearrangement. A good follow-up would be to revisit the interface
between the `NodeKey` and the node types to not rely on `pub(super)` in
this way.
- The `run_node` methods on each node type have been marked as
`pub(super)` to they can be called by `NodeKey`.
- Conversions to `NodeOutput` (other than `crate::python::Value` have
been moved to the submodule which actually relies on the conversion.